edit providers.py

+++ providers.py
@@ -8,7 +8,7 @@ with open("test.txt", "r") as f:
     b = f.read().split(" ")
 
 for i in range(0, n):
-    a[i] = int(b[randint(0, len(b))])
+    a[i] = float(b[randint(0, len(b))])
 
 
 print(a)
